+++ tests/testing_utils/mock_config.py
@@ -1,0 +1,29 @@
+# Copyright (c) 2024 Snowflake Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from unittest import mock
+
+from snowflake.cli.api.config import get_config_value
+
+
+def mock_config_key(key, value):
+    def get_config_side_effect_func(*args, **kwargs):
+        if kwargs["key"] == key:
+            return value
+        return get_config_value(*args, **kwargs)
+
+    return mock.patch(
+        "snowflake.cli.api.config.get_config_value",
+        side_effect=get_config_side_effect_func,
+    )
